test(suite-desktop): split and rename tests
